Changes to simplify and re-organize parameters setting. We now have one class for parameters (CbcParam) and one class for holding parameters (CbcSettings).
