+++ src/Attic/CbcModelParam.cpp
@@ -23,10 +23,7 @@
 #include "CbcModel.hpp"
 
 #include "CbcModelParam.hpp"
-#include "CbcSolverParam.hpp"
-#include "CbcSolverSettings.hpp"
-
-namespace {}
+#include "CbcParam.hpp"
 
 /*
   Constructors and destructors
@@ -158,185 +155,6 @@ CbcModelParam &CbcModelParam::operator=(const CbcModelParam &rhs) {
 */
 CbcModelParam::~CbcModelParam() { /* Nothing more to do */
 }
-
-//###########################################################################
-//###########################################################################
-
-namespace CbcModelParamUtils {
-
-//###########################################################################
-//###########################################################################
-
-/*
-  Set CbcModel defaults appropriate for cbc-generic.
-*/
-
-void setCbcModelDefaults(CbcModel *model)
-{
-  model->setIntParam(CbcModel::CbcMaxNumNode, (COIN_INT_MAX / 2));
-  model->setIntParam(CbcModel::CbcMaxNumSol, 999999);
-  model->setIntParam(CbcModel::CbcFathomDiscipline, 0);
-
-  model->setDblParam(CbcModel::CbcIntegerTolerance, 1.0e-6);
-  model->setDblParam(CbcModel::CbcInfeasibilityWeight, 0.0);
-  model->setDblParam(CbcModel::CbcCutoffIncrement, 1.0e-5);
-  model->setDblParam(CbcModel::CbcAllowableGap, 1.0e-10);
-  model->setDblParam(CbcModel::CbcAllowableFractionGap, 0.0);
-  // One year is 60x60x24x365 = 31,536,000 seconds.
-  model->setDblParam(CbcModel::CbcMaximumSeconds, 3.0e7);
-  model->setDblParam(CbcModel::CbcCurrentCutoff, 1.0e100);
-  model->setDblParam(CbcModel::CbcOptimizationDirection, 1.0);
-  model->setDblParam(CbcModel::CbcCurrentObjectiveValue, 1.0e100);
-  model->setDblParam(CbcModel::CbcCurrentMinimizationObjectiveValue, 1.0e100);
-  model->setDblParam(CbcModel::CbcStartSeconds, 0.0);
-
-  model->setNumberBeforeTrust(5);
-  model->setNumberStrong(5);
-
-  return;
-}
-
-//###########################################################################
-//###########################################################################
-
-/*
-  Function to push a double parameter.
-*/
-
-int pushCbcModelDblParam(CoinParam &param)
-
-{
-  CbcModelParam &cbcParam = dynamic_cast<CbcModelParam &>(param);
-  CbcModel *model = cbcParam.obj();
-  double val = cbcParam.dblVal();
-  CbcModelParamCode code = cbcParam.paramCode();
-
-  assert(model != 0);
-
-  int retval = 0;
-  /*
-      Translate the parameter code from CbcModelParamCode into the correct key
-     for CbcDblParam.
-    */
-  CbcModel::CbcDblParam key;
-  switch (code) {
-  case INTEGERTOLERANCE: {
-    key = CbcModel::CbcIntegerTolerance;
-    break;
-  }
-  case INFEASIBILITYWEIGHT: {
-    key = CbcModel::CbcInfeasibilityWeight;
-    break;
-  }
-  case INCREMENT: {
-    key = CbcModel::CbcCutoffIncrement;
-    break;
-  }
-  case ALLOWABLEGAP: {
-    key = CbcModel::CbcAllowableGap;
-    break;
-  }
-  case GAPRATIO: {
-    key = CbcModel::CbcAllowableFractionGap;
-    break;
-  }
-  case MAXSECONDSNOTIMPROVING: {
-    model->setMaxSecondsNotImproving(val);
-    break;
-  }
-  case TIMELIMIT: {
-    key = CbcModel::CbcMaximumSeconds;
-    break;
-  }
-  case CUTOFF: {
-    key = CbcModel::CbcCurrentCutoff;
-    break;
-  }
-  default: {
-    std::cerr << "pushCbcModelDblParam: no equivalent CbcDblParam for "
-              << "parameter code `" << code << "'." << std::endl;
-    retval = -1;
-    break;
-  }
-  }
-
-  bool result = model->setDblParam(key, val);
-  if (result == false) {
-    retval = -1;
-  }
-
-  return (retval);
-}
-
-//###########################################################################
-//###########################################################################
-
-/*
-  Function to push an integer parameter.
-*/
-
-int pushCbcModelIntParam(CoinParam &param)
-
-{
-  CbcModelParam &cbcParam = dynamic_cast<CbcModelParam &>(param);
-  CbcModel *model = cbcParam.obj();
-  int val = cbcParam.intVal();
-  CbcModelParamCode code = cbcParam.paramCode();
-
-  assert(model != 0);
-
-  int retval = 0;
-  /*
-      Translate the parameter code from CbcModelParamCode into the correct key
-     for CbcIntParam, or call the appropriate method directly.
-    */
-  CbcModel::CbcIntParam key = CbcModel::CbcLastIntParam;
-  switch (code) {
-  case CUTPASS: {
-    model->setMaximumCutPassesAtRoot(val);
-    break;
-  }
-  case LOGLEVEL: {
-    CoinMessageHandler *hndl = model->messageHandler();
-    assert(hndl != 0);
-    hndl->setLogLevel(val);
-    break;
-  }
-  case MAXNODESNOTIMPROVING: {
-    model->setMaxNodesNotImproving(val);
-    break;
-  }
-  case MAXSOLS: {
-    model->setMaxSolutions(val);
-    break;
-  }
-  case NUMBERBEFORE: {
-    model->setNumberBeforeTrust(val);
-    break;
-  }
-  default: {
-    std::cerr << "pushCbcModelIntParam: no equivalent CbcIntParam for "
-              << "parameter code `" << code << "'." << std::endl;
-    retval = -1;
-    break;
-  }
-  }
-
-  if (key != CbcModel::CbcLastIntParam) {
-    bool result = model->setIntParam(key, val);
-    if (result == false) {
-      retval = -1;
-    }
-  }
-
-  return (retval);
-}
-
-} // end namespace CbcModelParamUtils
-
-/* vi: softtabstop=2 shiftwidth=2 expandtab tabstop=2
- */
-
 
 /* vi: softtabstop=2 shiftwidth=2 expandtab tabstop=2
  */
