+++ src/CbcBab.cpp
@@ -23,9 +23,9 @@
 #include "CbcModel.hpp"
 #include "CbcStrategy.hpp"
 
-#include "CbcModelParam.hpp"
-#include "CbcSolverParam.hpp"
-#include "CbcSolverSettings.hpp"
+#include "CbcParam.hpp"
+#include "CbcParamUtils.hpp"
+#include "CbcSettings.hpp"
 
 #define CBC_TRACK_SOLVERS 1
 // #define COIN_CBC_VERBOSITY 5
@@ -159,7 +159,7 @@ bool solveRelaxation(CbcModel *model)
   Helper routine to establish a priority vector.
 */
 
-void setupPriorities(CbcModel *model, CbcSolverParam::BPMode how)
+void setupPriorities(CbcModel *model, CbcSettings::BPMode how)
 
 {
   int numCols = model->getNumCols();
@@ -174,9 +174,9 @@ void setupPriorities(CbcModel *model, CbcSolverParam::BPMode how)
   for (iColumn = 0; iColumn < numCols; iColumn++) {
     if (model->isInteger(iColumn)) {
       sort[n] = n;
-      if (how == CbcSolverParam::BPCost) {
+      if (how == CbcSettings::BPCost) {
         dsort[n++] = -objective[iColumn];
-      } else if (how == CbcSolverParam::BPOrder) {
+      } else if (how == CbcSettings::BPOrder) {
         dsort[n++] = iColumn;
       } else {
         std::cerr << "setupPriorities: Unrecognised priority specification."
@@ -216,33 +216,33 @@ void setupPriorities(CbcModel *model, CbcSolverParam::BPMode how)
   enabled.
 */
 
-void installHeuristics(CbcSolverSettings *cbcSettings, CbcModel *model)
+void installHeuristics(CbcSettings *cbcSettings, CbcModel *model)
 
 {
-  CbcSolverParam::HeurMode mode;
+  CbcSettings::HeurMode mode;
   CbcHeuristic *gen;
   CbcTreeLocal *localTree;
   /*
       FeasPump goes first because it only works before there's a solution.
     */
   mode = cbcSettings->getFeasPump(gen, model);
-  if (mode != CbcSolverParam::HeurOff) {
+  if (mode != CbcSettings::HeurOff) {
     model->addHeuristic(gen, "FeasPump");
   }
   mode = cbcSettings->getRounding(gen, model);
-  if (mode != CbcSolverParam::HeurOff) {
+  if (mode != CbcSettings::HeurOff) {
     model->addHeuristic(gen, "Rounding");
   }
   mode = cbcSettings->getCombine(gen, model);
-  if (mode != CbcSolverParam::HeurOff) {
+  if (mode != CbcSettings::HeurOff) {
     model->addHeuristic(gen, "Combine");
   }
   mode = cbcSettings->getGreedyCover(gen, model);
-  if (mode != CbcSolverParam::HeurOff) {
+  if (mode != CbcSettings::HeurOff) {
     model->addHeuristic(gen, "GCov");
   }
   mode = cbcSettings->getGreedyEquality(gen, model);
-  if (mode != CbcSolverParam::HeurOff) {
+  if (mode != CbcSettings::HeurOff) {
     model->addHeuristic(gen, "GEq");
   }
   /*
@@ -250,7 +250,7 @@ void installHeuristics(CbcSolverSettings *cbcSettings, CbcModel *model)
      the model.
     */
   mode = cbcSettings->getLocalTree(localTree, model);
-  if (mode != CbcSolverParam::HeurOff) {
+  if (mode != CbcSettings::HeurOff) {
     model->passInTreeHandler(*localTree);
   }
 
@@ -264,12 +264,12 @@ void installHeuristics(CbcSolverSettings *cbcSettings, CbcModel *model)
   figure out stored cuts.
 */
 
-void installCutGenerators(CbcSolverSettings *cbcSettings, CbcModel *model)
+void installCutGenerators(CbcSettings *cbcSettings, CbcModel *model)
 
 {
   int switches[20];
   int genCnt = 0;
-  CbcSolverParam::CGMode mode;
+  CbcSettings::CGMode mode;
   CglCutGenerator *gen;
 
   /*
@@ -281,21 +281,21 @@ void installCutGenerators(CbcSolverSettings *cbcSettings, CbcModel *model)
       it'll carry on regardless.
     */
 
-  int howOften[CbcSolverParam::CGEndMarker];
-  howOften[CbcSolverParam::CGOff] = -100;
-  howOften[CbcSolverParam::CGOn] = -1;
-  howOften[CbcSolverParam::CGRoot] = -99;
-  howOften[CbcSolverParam::CGIfMove] = -98;
-  howOften[CbcSolverParam::CGForceOn] = 1;
-  howOften[CbcSolverParam::CGForceOnBut] = 1;
+  int howOften[CbcSettings::CGEndMarker];
+  howOften[CbcSettings::CGOff] = -100;
+  howOften[CbcSettings::CGOn] = -1;
+  howOften[CbcSettings::CGRoot] = -99;
+  howOften[CbcSettings::CGIfMove] = -98;
+  howOften[CbcSettings::CGForceOn] = 1;
+  howOften[CbcSettings::CGForceOnBut] = 1;
 
   /*
       A negative value for rowCuts means that the specified modes happen only at
       the root.
     */
   mode = cbcSettings->getProbing(gen);
-  if (mode != CbcSolverParam::CGOff) {
-    if (mode == CbcSolverParam::CGForceOnBut) {
+  if (mode != CbcSettings::CGOff) {
+    if (mode == CbcSettings::CGForceOnBut) {
       CglProbing *probingGen = dynamic_cast<CglProbing *>(gen);
       probingGen->setRowCuts(-3);
     }
@@ -303,37 +303,37 @@ void installCutGenerators(CbcSolverSettings *cbcSettings, CbcModel *model)
     switches[genCnt++] = 0;
   }
   mode = cbcSettings->getGomory(gen);
-  if (mode != CbcSolverParam::CGOff) {
+  if (mode != CbcSettings::CGOff) {
     model->addCutGenerator(gen, howOften[mode], "Gomory");
     switches[genCnt++] = -1;
   }
   mode = cbcSettings->getKnapsack(gen);
-  if (mode != CbcSolverParam::CGOff) {
+  if (mode != CbcSettings::CGOff) {
     model->addCutGenerator(gen, howOften[mode], "Knapsack");
     switches[genCnt++] = 0;
   }
   mode = cbcSettings->getRedSplit(gen);
-  if (mode != CbcSolverParam::CGOff) {
+  if (mode != CbcSettings::CGOff) {
     model->addCutGenerator(gen, howOften[mode], "RedSplit");
     switches[genCnt++] = 1;
   }
   mode = cbcSettings->getClique(gen);
-  if (mode != CbcSolverParam::CGOff) {
+  if (mode != CbcSettings::CGOff) {
     model->addCutGenerator(gen, howOften[mode], "Clique");
     switches[genCnt++] = 0;
   }
   mode = cbcSettings->getMir(gen);
-  if (mode != CbcSolverParam::CGOff) {
+  if (mode != CbcSettings::CGOff) {
     model->addCutGenerator(gen, howOften[mode], "MIR2");
     switches[genCnt++] = -1;
   }
   mode = cbcSettings->getFlow(gen);
-  if (mode != CbcSolverParam::CGOff) {
+  if (mode != CbcSettings::CGOff) {
     model->addCutGenerator(gen, howOften[mode], "Flow");
     switches[genCnt++] = 1;
   }
   mode = cbcSettings->getTwomir(gen);
-  if (mode != CbcSolverParam::CGOff) {
+  if (mode != CbcSettings::CGOff) {
     model->addCutGenerator(gen, howOften[mode], "2-MIR");
     switches[genCnt++] = 1;
   }
@@ -466,7 +466,7 @@ void setupObjects(OsiSolverInterface *osi, bool didIPP, CglPreProcess *ippObj)
 
 } // namespace
 
-namespace CbcSolverParamUtils {
+namespace CbcParamUtils {
 
 /*
   Run branch-and-cut.
@@ -475,8 +475,8 @@ namespace CbcSolverParamUtils {
 int doBaCParam(CoinParam &param)
 
 {
-  CbcSolverParam &cbcParam = dynamic_cast<CbcSolverParam &>(param);
-  CbcSolverSettings *cbcSettings = cbcParam.obj();
+  CbcParam &cbcParam = dynamic_cast<CbcParam &>(param);
+  CbcSettings *cbcSettings = cbcParam.settings();
   assert(cbcSettings != 0);
   CbcModel *model = cbcSettings->getModel();
   assert(model != 0);
@@ -489,9 +489,9 @@ int doBaCParam(CoinParam &param)
   } else {
     retval = -1;
   }
-  cbcSettings->setBaBStatus(CbcSolverParam::BACAbandon,
-                            CbcSolverParam::BACmInvalid,
-                            CbcSolverParam::BACwNotStarted, false, 0);
+  cbcSettings->setBaBStatus(CbcSettings::BACAbandon,
+                            CbcSettings::BACmInvalid,
+                            CbcSettings::BACwNotStarted, false, 0);
   /*
       We ain't gonna do squat without a good model.
     */
@@ -520,7 +520,7 @@ int doBaCParam(CoinParam &param)
       Solve the root relaxation. Bail unless it solves to optimality.
     */
   if (!solveRelaxation(&babModel)) {
-    cbcSettings->setBaBStatus(CbcSolverParam::BACwBareRoot);
+    cbcSettings->setBaBStatus(CbcSettings::BACwBareRoot);
     return (0);
   }
 #if COIN_CBC_VERBOSITY > 0
@@ -530,7 +530,7 @@ int doBaCParam(CoinParam &param)
   /*
       Are we up for fixing variables based on reduced cost alone?
     */
-  if (cbcSettings->getDjFixMode() == CbcSolverParam::ParamOff) {
+  if (cbcSettings->getDjFixMode() == CbcSettings::ParamOff) {
     reducedCostHack(babSolver, cbcSettings->getDjFixThreshold());
   }
   /*
@@ -550,10 +550,10 @@ int doBaCParam(CoinParam &param)
 
   int numberChanged = 0;
   int numberOriginalColumns = babSolver->getNumCols();
-  CbcSolverParam::IPPMode ippMode = cbcSettings->getIPPMode();
+  CbcSettings::IPPMode ippMode = cbcSettings->getIPPMode();
 
-  if (!(ippMode == CbcSolverParam::IPPOff ||
-        ippMode == CbcSolverParam::IPPStrategy)) {
+  if (!(ippMode == CbcSettings::IPPOff ||
+        ippMode == CbcSettings::IPPStrategy)) {
     double timeLeft = babModel.getMaximumSeconds();
     preIppSolver = babSolver->clone();
     OsiSolverInterface *ippSolver;
@@ -581,13 +581,13 @@ int doBaCParam(CoinParam &param)
           respectively.
         */
     int convert = 0;
-    if (ippMode == CbcSolverParam::IPPEqual) {
+    if (ippMode == CbcSettings::IPPEqual) {
       convert = -1;
-    } else if (ippMode == CbcSolverParam::IPPEqualAll) {
+    } else if (ippMode == CbcSettings::IPPEqualAll) {
       convert = -2;
-    } else if (ippMode == CbcSolverParam::IPPSOS) {
+    } else if (ippMode == CbcSettings::IPPSOS) {
       convert = 2;
-    } else if (ippMode == CbcSolverParam::IPPTrySOS) {
+    } else if (ippMode == CbcSettings::IPPTrySOS) {
       convert = 3;
     }
 
@@ -609,7 +609,7 @@ int doBaCParam(CoinParam &param)
       std::cout << "Integer preprocess says infeasible or unbounded"
                 << std::endl;
       delete preIppSolver;
-      cbcSettings->setBaBStatus(CbcSolverParam::BACwIPP);
+      cbcSettings->setBaBStatus(CbcSettings::BACwIPP);
       return (0);
     }
 #if COIN_CBC_VERBOSITY > 0
@@ -624,7 +624,7 @@ int doBaCParam(CoinParam &param)
     preIppSolver->setHintParam(OsiDoInBranchAndCut, false, OsiHintDo);
     ippSolver->setHintParam(OsiDoInBranchAndCut, false, OsiHintDo);
 
-    if (ippMode == CbcSolverParam::IPPSave) {
+    if (ippMode == CbcSettings::IPPSave) {
       ippSolver->writeMps("presolved", "mps", 1.0);
       std::cout << "Integer preprocessed model written to `presolved.mps' "
                 << "as minimisation problem." << std::endl;
@@ -640,7 +640,7 @@ int doBaCParam(CoinParam &param)
 #endif
     if (!solveRelaxation(&babModel)) {
       delete preIppSolver;
-      cbcSettings->setBaBStatus(CbcSolverParam::BACwIPPRelax);
+      cbcSettings->setBaBStatus(CbcSettings::BACwIPPRelax);
       return (0);
     }
 #if COIN_CBC_VERBOSITY > 0
@@ -658,7 +658,7 @@ int doBaCParam(CoinParam &param)
       If we're using the COSTSTRATEGY option, set up priorities here and pass
       them to the babModel.
     */
-  if (cbcSettings->getBranchingPriorityMode() != CbcSolverParam::BPOff) {
+  if (cbcSettings->getBranchingPriorityMode() != CbcSettings::BPOff) {
      setupPriorities(&babModel, cbcSettings->getBranchingPriorityMode());
   }
   /*
@@ -746,7 +746,7 @@ int doBaCParam(CoinParam &param)
       accidentally equivalent to IPPOn.
     */
 
-  if (ippMode == CbcSolverParam::IPPStrategy) {
+  if (ippMode == CbcSettings::IPPStrategy) {
     CbcStrategyDefault strategy(true, 5, 5);
     strategy.setupPreProcessing(1);
     babModel.setStrategy(strategy);
@@ -777,7 +777,7 @@ int doBaCParam(CoinParam &param)
     */
   if (cbcSettings->getDebugCreate() == "createAfterPre" &&
       babModel.bestSolution()) {
-    CbcSolverParamUtils::saveSolution(babSolver, "debug.file");
+    CbcParamUtils::saveSolution(babSolver, "debug.file");
   }
   /*
       Print some information about branch-and-cut.
@@ -806,7 +806,7 @@ int doBaCParam(CoinParam &param)
   /*
       If we performed integer preprocessing, time to back it out.
     */
-  if (ippMode != CbcSolverParam::IPPOff) {
+  if (ippMode != CbcSettings::IPPOff) {
 #if CBC_TRACK_SOLVERS > 0
     std::cout << "doBaCParam: solver passed to IPP postprocess is " << std::hex
               << babSolver << std::dec << "." << std::endl;
@@ -823,7 +823,7 @@ int doBaCParam(CoinParam &param)
       Write out postprocessed solution to debug file, if requested.
     */
   if (cbcSettings->getDebugCreate() == "create" && babModel.bestSolution()) {
-    CbcSolverParamUtils::saveSolution(babSolver, "debug.file");
+    CbcParamUtils::saveSolution(babSolver, "debug.file");
   }
   /*
       If we have a good solution, detach the solver with the answer. Fill in the
@@ -836,7 +836,7 @@ int doBaCParam(CoinParam &param)
     keepAnswerSolver = true;
     answerSolver = babSolver;
   }
-  cbcSettings->setBaBStatus(CbcSolverParam::BACwBAC,
+  cbcSettings->setBaBStatus(CbcSettings::BACwBAC,
                             keepAnswerSolver, answerSolver);
   /*
       And one last bit of information & statistics.
@@ -853,7 +853,7 @@ int doBaCParam(CoinParam &param)
   return (0);
 }
 
-} // namespace CbcSolverParamUtils
+} // namespace CbcParamUtils
 
 /* vi: softtabstop=2 shiftwidth=2 expandtab tabstop=2
  */
