+++ src/unitTestClp.cpp
@@ -128,7 +128,7 @@ int CbcClpUnitTest(const CbcModel &saveModel, const std::string &dirMiplib,
 		   int testSwitch, const double *stuff, int argc,
 		   const char ** argv,
 		   int callBack(CbcModel *currentSolver, int whereFrom),
-		   CbcSolverSettings &cbcSettings)
+		   CbcSettings &cbcSettings)
 {
   // Stop Windows popup
   WindowsErrorPopupBlocker();
